Renamed command folder to commands
