+++ src/main/java/duke/commands/CommandMark.java
@@ -1,4 +1,4 @@
-package duke.command;
+package duke.commands;
 
 import duke.Ui;
 import duke.exceptions.DukeCeption;
